Initial implementation of a new lockless hashtable mpmc (#262)
This PR implements a new lockless multi producer and multi consumer
parallel hashtable which is capable of achieving amazing numbers
(included at the end of the summary)!

This new hashtable works in a very different way from the previous one
although the goal keeps being the same: spread the contention of the
threads as much as possible and reduce the amount of times the data in
memory are changed.

The old approach was relying on a combination of memory fencing and
userspace spinlocks to perform lockless GET operations meanwhile using
userspace spinlocks for the SET and DEL operations with a lock every 14
buckets, the results were great as an hashtable with 10000 was being
controlled by about 714 locks, but the downside was a lot of memory
changes, also impacting the cache of the cores, done by 1 single thread
to write a value.

The new approach relies on a combination of approaches and components:
- an epoch operation queue, each time an operation is triggered with the
hashtable an operation is pushed to the queue and marked as completed
once done, guaranteeing that the data read by the thread after pushing
the new operation to the queue will never be deleted until the operation
itself is completed
- an epoch garbage collector, which operates in combination with the
epoch operation queue, every time a bucket is deleted the associated key
value data are not deleted immediately but insted staged for deletion,
will only be deleted once all the operation started before the staging
are marked as completed
- a 3 pass approach to insert new buckets which guarantees that at some
point if multiple threads are fighting to insert the same key one will
be successful
- which becomes 4 pass in case an insertion is carried out during an
upsize if the key is new
- the key value data are not stored inside the hashtable itself but as a
pointer to an external structure, this makes a bucket much larger (16
bytes) but drammatically speed up the copy time and makes possible to
use 128 bit atomic operations to update a bucket
- pointer tagging to store status in the pointer of the key value itself
- a transaction spinlock, as it's in use in the current hashtable, for
the single/multi key transactions and the Read-Modify-Write operations

The hashtable will implement some extra optimizations to box the upper
level data (the storagedb_entry_index) into the lower level data (the
key value itself) using a combination of an union and the code in the
header using some defines to create a typed version of the hashtable.

This new hashtable will provide much better performances when batching
operations or operating in a cluster.

The PR includes new tests and new benchmarks, used to generate the
numbers initially mentioned.

The PR is drammatically large:
- 74 commits
- 18 files changed
- 4465 new lines
- only 71 deletions

Talking about numbers, here some

Operation | Threads | V1 - Million Op/s |  V2 - Million Op/s
-- | -- | -- | --
INSERT | 1 | 3.22191 | 5.02507
INSERT | 2 | 6.26603 | 9.74552
INSERT | 4 | 11.73655 | 19.20719
INSERT | 8 | 24.03989 | 38.92227
INSERT | 16 | 42.58111 | 72.36672
INSERT | 32 | 69.45982 | 129.38364
INSERT | 64 | 109.68172 | **197.67271**
UPDATE | 1 | 3.18695 | 5.55167
UPDATE | 2 | 6.16648 | 11.06459
UPDATE | 4 | 11.53689 | 21.79368
UPDATE | 8 | 22.8819 | 43.98683
UPDATE | 16 | 37.42774 | 83.61501
UPDATE | 32 | 63.96726 | 143.30408
UPDATE | 64 | 89.5212 | **236.87745**

The V2 hashtable is between **2 and 2.5 times faster** than the V1, the
current implementation.

The hardware used for benchmarking was an EPYC 7502P (32 core, 64 hw
thread, default bios settings) with 256GB RDIMM DDR4 3200mhz.

Closes #181

+++ src/data_structures/hashtable/spsc/hashtable_spsc.c
@@ -22,9 +22,9 @@ hashtable_spsc_t *hashtable_spsc_new(
 
     assert(buckets_count > 0);
     assert(max_range > 0);
-    assert(buckets_count < UINT16_MAX);
+    assert(buckets_count < UINT32_MAX);
 
-    buckets_count_pow2_next = (uint16_t)pow2_next(buckets_count);
+    buckets_count_pow2_next = (uint32_t)pow2_next(buckets_count);
     buckets_to_allocate = buckets_count_pow2_next + max_range;
 
     assert(buckets_count_pow2_next >= buckets_count);
